contextとtypesの整理
