+++ components/molecules/StackList.tsx
@@ -3,7 +3,8 @@ import StackCard from '@/components/molecules/StackCard';
 import SelectBox from './SelectBox';
 import { SelectChangeEvent } from '@mui/material/Select';
 import { getSession } from '@/utiliry/session';
-import { ApiOptions, StackProps } from '@/types/types';
+import { ApiOptions } from '@/types/api';
+import { StackProps } from '@/types/stack';
 import axios from 'axios';
 
 const StackList: FC = () => {
