+++ pages/timeline/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { NextPage } from 'next'
 import Layout from '@/components/organisms/Layout'
 import StackList from '@/components/molecules/StackList'
