+++ types/skill.ts
@@ -1,0 +1,6 @@
+export type SkillProps = {
+  id: number;
+  name: string;
+  created_at?: string;
+  updated_at?: string;
+};
