+++ types/team.ts
@@ -1,0 +1,6 @@
+export type TeamProps = {
+  id: number;
+  name: string;
+  created_at: string;
+  updated_at: string;
+}
